feat: Display NIP-01 metadata in the timeline
This change enhances the timeline to display NIP-01 metadata, such as usernames, instead of just public keys.

- Fetches NIP-01 metadata for users in the timeline.
- Creates a `TimelinePost` struct to hold both the status and the author's metadata.
- Updates the UI to display the username in a card-based layout for each post.
- Refactors the codebase by splitting UI and Nostr logic into separate modules (`ui.rs` and `nostr_client.rs`) for better maintainability.
